Create latestVersion.txt
unity tells my game to go to github and check the latest version. if the latest version is equal to the current version that the game is running on, then unity ignores it. However if the latest version is greater than the current version that the game is running on, the game forces you to update or quit the game.